Consolidate ScanNext & ScanZone components
- Consolidate ScanNext and ScanZone components into one reusable component
- Cleanup backend code to use more general variables
-
